refactor: move recorder input preferences to lib
